+++ src/carto/constants.js
@@ -11,6 +11,3 @@ export const cartoHousingActionsTable = 'public.emergency_housing_actions';
 export const cartoCountiesTable = 'public.us_county_boundaries';
 export const cartoStatesTable = 'public.states_and_provinces_global';
 export const cartoNationsTable = 'public.countries';
-
-export const bingApiKey =
-  'Al2-1GXrd8GzwtSAID3J3LJn-flLLgNWzNtsT5nnSKW8dA2ClgaXXXMQR6WfE6wE';
